requirements.txt
# Core data processing dependencies
pandas>=2.3.0
numpy>=1.26.0
scikit-learn>=1.7.0

# Visualization
matplotlib>=3.10.0
seaborn>=0.12.0

# Configuration management
pyyaml>=6.0.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.18.0
ruff>=0.1.15

# Jupyter and notebook support
jupyter>=1.1.0
ipykernel>=6.30.0

# Optional: Additional analysis tools
scipy>=1.16.0